custom_storages.py
from storages.backends.s3boto3 import S3Boto3Storage
from jertap_backend import settings


class StaticStorage(S3Boto3Storage):
    location = settings.STATICFILES_LOCATION


class MediaStorage(S3Boto3Storage):
    location = settings.MEDIAFILES_LOCATION
    file_overwrite = False
