requirements.txt
Django==4.2.7
gunicorn==21.2.0
djangorestframework==3.14.0
django-filter==23.3
python-decouple==3.8
dj-database-url==2.1.0
django-cors-headers==4.3.0
django-dirtyfields==1.9.2
pillow==10.1.0
coreapi==2.3.3
djangorestframework-simplejwt==5.3.0
django-storages==1.14.2
boto3==1.34.11
botocore==1.34.11
django-cleanup==8.0.0
django-rest-auth==0.9.5
psycopg2-binary==2.9.9
google-api-core==2.13.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
facebook-sdk==3.1.0
pandas==2.1.4
celery==5.3.6
requests==2.31.0
pycurl==7.45.2
django-ses==3.5.2